added SQLite instead of IMDb, made initializer, made DTO controllers (used to avoid all fields to be shown in API output)

+++ FMF_Backend/Models/Order.cs
@@ -8,7 +8,7 @@ public class Order {
 
     public Order(){}
 
-    public Order(GroceryList groceryList, Driver driver, decimal finalCheck){
+    public Order(GroceryList groceryList, Driver driver, double finalCheck){
 
         groceryList = GroceryList;
         driver = Driver;
@@ -21,11 +21,9 @@ public class Order {
     public Driver Driver { get; set; }
     public GroceryList GroceryList { get; set; }
     public Customer Customer { get; set; }
-    public decimal FinalCheck { get; set; }
-    
+    public double FinalCheck { get; set; }
     [DataType(DataType.Date)]
     public DateTime OrderTime { get; set; }
-
     public List<OrderLine> OrderLines { get; set; }
