+++ FMF_Backend/Models/OrderLine.cs
@@ -1,11 +1,8 @@
-namespace FMF_Backend.Models
-{
-    public class OrderLine
-    {
+namespace FMF_Backend.Models{
+    public class OrderLine{
         public OrderLine() {}
 
-        public OrderLine(Product product, int amount)
-        {
+        public OrderLine(Product product, double amount){
             Product = product;
             Amount = amount;
         }
@@ -13,6 +10,6 @@ namespace FMF_Backend.Models
         public int Id { get; set; }
         public int ProductId { get; set; }
         public Product Product { get; set; }
-        public int Amount { get; set; }
+        public double Amount { get; set; }
     }
 }
