+++ FMF_Backend/Models/Product.cs
@@ -6,7 +6,8 @@ public class Product {
 
     public Product() {}
 
-    public Product(string productName, string supplier, decimal priceStore1, decimal priceStore2, decimal priceFMF){
+    public Product(string productName, string supplier, 
+    double priceStore1, double priceStore2, double priceFMF){
         productName = ProductName;
         supplier = Supplier;
         priceStore1 = PriceStore1;
@@ -19,13 +20,13 @@ public class Product {
     public string ProductName { get; set; }
 
     [DataType(DataType.Currency)]
-    public decimal PriceStore1 { get; set; }
+    public double PriceStore1 { get; set; }
 
     [DataType(DataType.Currency)]   
-    public decimal PriceStore2 { get; set; }
+    public double PriceStore2 { get; set; }
 
     [DataType(DataType.Currency)]
-    public decimal PriceFMF { get; set; }
+    public double PriceFMF { get; set; }
 
     public string Supplier { get; set; }
